+++ packages/vendure-plugin-google-cloud-tasks/CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.1.2 (2023-09-26)
 
-- Added custom error handler to inspect errors from failed jobs([#262](https://github.com/Pinelab-studio/pinelab-vendure-plugins/pull/262))
+- Added `onJobFailure` option to inspect errors from failed jobs([#262](https://github.com/Pinelab-studio/pinelab-vendure-plugins/pull/262))